requirements.txt
simpleaudio
pyyaml
